Load audio files concurrently
